+++ assets/javascripts/kitten/components/search/search-input/stories.js
@@ -1,0 +1,29 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import { withKnobs } from '@storybook/addon-knobs'
+import { SearchInput } from './index'
+import { Marger } from '../../layout/marger'
+import { Container } from '../../grid/container'
+import { Grid, GridCol } from '../../grid/grid'
+
+const StoryGrid = ({ children }) => (
+  <Container>
+    <Grid>
+      <GridCol col="3">
+        <Marger top="5" bottom="5">
+          {children}
+        </Marger>
+      </GridCol>
+    </Grid>
+  </Container>
+)
+
+storiesOf('Search/SearchInput', module)
+  .addDecorator(withKnobs)
+  .add('default', () => {
+    return (
+      <StoryGrid>
+        <SearchInput placeholder="Votre recherche" />
+      </StoryGrid>
+    )
+  })