+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Fix: Remove margin when `title` prop is empty.
+- Fix: Remove margin when `title` prop is empty in `CounterBlock` component.
 
 ## [1.4.1] - 2019-01-10