+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add `activePaginationColor`, `paginationColor` and `paginationAlign`
   props on `SimpleCarousel`.
 - Fix: Adds the ability to edit and add styles on `BulletList`.
+- Fix: Remove margin when `title` prop is empty in `CounterBlock` component.
 
 ## [1.4.1] - 2019-01-10