+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: new `Overlay` component with `zIndex` (number) and `isActive` (bool) props.
 - Fix: Remove margin when `title` prop is empty in `CounterBlock` component.
 
 ## [1.4.1] - 2019-01-10