+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: animation on `Accordeon` component
+- Fix: Remove margin when `title` prop is empty in `CounterBlock` component.
 
 ## [1.4.1] - 2019-01-10