+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Remove margin when `title` prop is empty in `CounterBlock` component.
+- Fix: Force `type="button"` on `buttonQuestionMarkIcon` component.
 
 ## [1.4.1] - 2019-01-10