+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Feature: animation on `Accordeon` component
 - Fix: Remove margin when `title` prop is empty in `CounterBlock` component.
+- Fix: Force `type="button"` on `buttonQuestionMarkIcon` component.
 
 ## [1.4.1] - 2019-01-10