+++ CHANGELOG.md
@@ -4,9 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: animation on `Accordeon` component
-- Fix: Remove margin when `title` prop is empty in `CounterBlock` component.
-- Fix: Force `type="button"` on `buttonQuestionMarkIcon` component.
+## [1.5.0] - 2019-01-16
+
+Feature:
+- Animation on `Accordeon` component.
+
+Fixes:
+- Remove margin when `title` prop is empty in `CounterBlock` component.
+- Force `type="button"` on `buttonQuestionMarkIcon` component.
 
 ## [1.4.1] - 2019-01-10